chore: fix build
