+++ libs/wallet/src/wagmi/config.ts
@@ -13,7 +13,8 @@ import { getReownDefaultNetwork } from './getReownDefaultNetwork'
 
 import { bindActiveProvider } from '../bindActiveProvider'
 import { interceptEIP6963Providers } from '../providerIsolation'
-import { SAFE_CONNECTOR_ID, SUPPORTED_REOWN_NETWORKS } from '../reown/consts'
+import { SAFE_CONNECTOR_ID } from '../reown/consts'
+import { SUPPORTED_REOWN_NETWORKS } from '../reown/networks'
 import { connectWalletById } from '../utils/connectWalletById'
 import { getIsSafeAppIframe } from '../utils/getIsSafeAppIframe'
 import { wagmiStorage } from '../wagmiStorage'
